Remove spurious __init__ variables
When you execute

In [1]: import human_readable_project_name as he
In [2]: help(he)

The help page would previously show version_file and version_filename
in the DATA section.

Now, the version file reading is done in a private function and does
not show up in the help page anymore.